nicer control flow in draw_untextured and changed types of ouput width and height to size_t

+++ src/renderer.h
@@ -1,6 +1,7 @@
 #ifndef RENDERER_H
 #define RENDERER_H
 
+#include <stddef.h>
 #include <stdint.h>
 #include "map.h"
 #include "maths.h"
@@ -32,19 +33,19 @@ typedef struct {
 
 
 typedef void(*draw_function_t)(image_t*, tex_atlas_t*,  ray_cast_result_t*,
-                		int px, int py);
+                		size_t px, size_t py);
 // responsible for drawing a pixel at (px, py) into the image
 
 
 // px, py: position in the image
-void draw_untextured(image_t *img, tex_atlas_t *ta, ray_cast_result_t *rc_res, int px, int py);
+void draw_untextured(image_t *img, tex_atlas_t *ta, ray_cast_result_t *rc_res, size_t px, size_t py);
 
 
 void tex_atlas_load(tex_atlas_t* tex_atlas, FILE *f, size_t nx, size_t ny);
 
 // renders the map viewed from (px, py), with rotation and fov into img
 // using draw_function
-void render(FILE *out, map_t *map, int w, int h, double px, double py,
+void render(FILE *out, map_t *map, size_t w, size_t h, double px, double py,
 		double fov, double rotation, draw_function_t draw,
 		tex_atlas_t *tex_atlas, image_t *img);
 #endif
